// src/generated/model/customer.ts
/**
 * OpenAPI ZbW car rent
 * zbw car rent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Place } from './place';


/**
 * A customer
 */
export interface Customer { 
    idCustomer?: number;
    firstname?: string;
    lastname?: string;
    address?: string;
    addressNr?: string;
    place?: Place;
}